refactor(ui): split error overlay state helpers
